Release preparation (#327)
* Fixing state loading for AAX

* Fixing clap-validator preset load for factory presets

* Fixing issues with some of the new processors

* Fix original editor positions being off on Windows

* Fixing preset discovery for user presets provider when user presets folder has not been set up

* Remove dead comment

* Make MIDI Modulator and Param Modulator bipolar by default

* Tweak DC Bias range and process Level Detective even if output is not connected

* Revert Gate to before level-tracking changes

* Fix Xcode warning

* Apply clang-format

* Remove dead things

---------

+++ src/processors/drive/centaur/FeedForward2.h
@@ -89,8 +89,8 @@ struct FeedForward2WDF
 
     ResistorCapacitorParallelT<float> R5_C4 { 5.1e3f, 68e-9f };
     WDFSeriesT<float, decltype (R5_C4), decltype (S6)> S7 { R5_C4, S6 };
-    PolarityInverterT<double, decltype (S7)> I1 { S7 };
+    PolarityInverterT<float, decltype (S7)> I1 { S7 };
 
-    IdealVoltageSourceT<double, decltype (I1)> Vin { I1 };
+    IdealVoltageSourceT<float, decltype (I1)> Vin { I1 };
 };
 } // namespace gain_stage
